requirements.txt
deepl.py
srt
